RUI-000: introduce remarkable pro
