+++ embeddable.lifecycle.ts
@@ -1,5 +1,5 @@
-import { injectCssVariables } from './src/remarkable-ui-embeddables/theme/styles/styles.utils';
-import { Theme } from './src/remarkable-ui-embeddables/theme/theme.types';
+import { injectCssVariables } from './src/remarkable-pro/theme/styles/styles.utils';
+import { Theme } from './src/remarkable-pro/theme/theme.types';
 
 export default {
   onThemeUpdated: (theme: Theme) => injectCssVariables(theme.styles),
