+++ embeddable.theme.ts
@@ -1,5 +1,5 @@
-import { Theme } from './src/remarkable-ui-embeddables/theme/theme.types';
-import { remarkableTheme } from './src/remarkable-ui-embeddables/theme/theme.constants';
+import { Theme } from './src/remarkable-pro/theme/theme.types';
+import { remarkableTheme } from './src/remarkable-pro/theme/theme.constants';
 
 const themeProvider = (): Theme => {
   return remarkableTheme;
