+++ src/index.tsx
@@ -1,9 +1,9 @@
-export { defaultChartsMenuOptions } from './remarkable-ui-embeddables/theme/theme.constants';
-export * from './remarkable-ui-embeddables/theme/theme.types';
-export * from './remarkable-ui-embeddables/theme/theme.constants';
+export { defaultChartsMenuOptions } from './remarkable-pro/theme/theme.constants';
+export * from './remarkable-pro/theme/theme.types';
+export * from './remarkable-pro/theme/theme.constants';
 export type {
   NumberFormatter,
   DateTimeFormatter,
   StringFormatter,
   ThemeFormatter,
-} from './remarkable-ui-embeddables/theme/formatter/formatter.types';
+} from './remarkable-pro/theme/formatter/formatter.types';
