+++ src/remarkable-pro/components/charts/pies/DonutChartPro/index.tsx
@@ -1,24 +1,17 @@
 import { useTheme } from '@embeddable.com/react';
 import { Theme } from '../../../../theme/theme.types';
 import { DefaultPieChartOptions, getDefaultPieChartOptions, getPieChartData } from '../pies.utils';
-import { DefaultReadyMadePieChartProps } from '../pies.types';
-import { DataResponse, Measure } from '@embeddable.com/core';
-import { getThemeFormatter } from '../../../../theme/formatter/formatter.utils';
+import { DefaultPieChartProps } from '../pies.types';
 import { i18nSetup } from '../../../../theme/i18n/i18n';
 import { ChartCard } from '../../shared/ChartCard/ChartCard';
 import { DonutChart } from '../../../../../remarkable-ui';
 import { mergician } from 'mergician';
 import { resolveI18nProps } from '../../../component.utils';
 
-type ReadyMadeDonutLabelChartProps = DefaultReadyMadePieChartProps & {
-  innerLabelMeasure: Measure;
-  innerLabelText: string;
-  resultsInnerLabel: DataResponse;
-};
+type DonutChartProProps = DefaultPieChartProps;
 
-const ReadyMadeDonutChart = (props: ReadyMadeDonutLabelChartProps) => {
+const DonutChartPro = (props: DonutChartProProps) => {
   const theme = useTheme() as Theme;
-  const themeFormatter = getThemeFormatter(theme);
   i18nSetup(theme);
 
   const {
@@ -31,24 +24,10 @@ const ReadyMadeDonutChart = (props: ReadyMadeDonutLabelChartProps) => {
     showTooltips,
     showValueLabels,
     title,
-    innerLabelMeasure,
-    resultsInnerLabel,
-    innerLabelText,
     onSegmentClick,
   } = resolveI18nProps(props);
 
   const data = getPieChartData({ data: results.data, dimension, measure, maxLegendItems }, theme);
-
-  const handleSegmentClick = (index: number | undefined) => {
-    onSegmentClick({
-      dimensionValue: index === undefined ? undefined : results.data?.[index]?.[dimension.name],
-    });
-  };
-
-  const label = themeFormatter.data(
-    innerLabelMeasure,
-    resultsInnerLabel?.data?.[0]?.[innerLabelMeasure.name],
-  );
 
   const options = mergician(
     getDefaultPieChartOptions(
@@ -60,8 +39,14 @@ const ReadyMadeDonutChart = (props: ReadyMadeDonutLabelChartProps) => {
       } as DefaultPieChartOptions,
       theme,
     ),
-    theme.charts.donutLabelChartOverrides ?? {},
+    theme.charts.donutChartOverrides ?? {},
   );
+
+  const handleSegmentClick = (index: number | undefined) => {
+    onSegmentClick({
+      dimensionValue: index === undefined ? undefined : results.data?.[index]?.[dimension.name],
+    });
+  };
 
   return (
     <ChartCard
@@ -71,15 +56,9 @@ const ReadyMadeDonutChart = (props: ReadyMadeDonutLabelChartProps) => {
       subtitle={description}
       title={title}
     >
-      <DonutChart
-        label={label}
-        subLabel={innerLabelText}
-        data={data}
-        options={options}
-        onSegmentClick={handleSegmentClick}
-      />
+      <DonutChart data={data} options={options} onSegmentClick={handleSegmentClick} />
     </ChartCard>
   );
 };
 
-export default ReadyMadeDonutChart;
+export default DonutChartPro;
