+++ src/remarkable-pro/components/charts/pies/DonutLabelChartPro/DonutLabelChartPro.emb.ts
@@ -1,7 +1,7 @@
 import { Value, loadData } from '@embeddable.com/core';
 import { defineComponent, EmbeddedComponentMeta, Inputs } from '@embeddable.com/react';
 
-import ReadyMadePieChart from './index';
+import PieChart from './index';
 import {
   dimension,
   dataset,
@@ -15,13 +15,22 @@ import {
 } from '../../../component.constants';
 
 export const meta = {
-  name: 'PieChart',
-  label: 'Pie Chart',
+  name: 'DonutLabelChartPro',
+  label: 'Donut Label Chart',
   category: 'Pie Charts',
   inputs: [
     dataset,
     measure,
     dimension,
+    { ...measure, name: 'innerLabelMeasure', label: 'Inner Label Measure' },
+    {
+      name: 'innerLabelText',
+      type: 'string',
+      label: 'Inner Label Text',
+      description: 'Text to display inside the donut chart',
+      required: false,
+      category: 'Component Data',
+    },
     title,
     description,
     showLegend,
@@ -44,13 +53,17 @@ export const meta = {
   ],
 } as const satisfies EmbeddedComponentMeta;
 
-export default defineComponent(ReadyMadePieChart, meta, {
+export default defineComponent(PieChart, meta, {
   props: (inputs: Inputs<typeof meta>) => {
     return {
       ...inputs,
       results: loadData({
         from: inputs.dataset,
         select: [inputs.measure, inputs.dimension],
+      }),
+      resultsInnerLabel: loadData({
+        from: inputs.dataset,
+        select: [inputs.innerLabelMeasure],
       }),
     };
   },
