+++ src/remarkable-pro/components/charts/pies/DonutLabelChartPro/index.tsx
@@ -1,17 +1,24 @@
 import { useTheme } from '@embeddable.com/react';
 import { Theme } from '../../../../theme/theme.types';
 import { DefaultPieChartOptions, getDefaultPieChartOptions, getPieChartData } from '../pies.utils';
-import { DefaultReadyMadePieChartProps } from '../pies.types';
+import { DefaultPieChartProps } from '../pies.types';
+import { DataResponse, Measure } from '@embeddable.com/core';
+import { getThemeFormatter } from '../../../../theme/formatter/formatter.utils';
 import { i18nSetup } from '../../../../theme/i18n/i18n';
 import { ChartCard } from '../../shared/ChartCard/ChartCard';
 import { DonutChart } from '../../../../../remarkable-ui';
 import { mergician } from 'mergician';
 import { resolveI18nProps } from '../../../component.utils';
 
-type ReadyMadeDonutChartProps = DefaultReadyMadePieChartProps;
+type DonutLabelChartProProps = DefaultPieChartProps & {
+  innerLabelMeasure: Measure;
+  innerLabelText: string;
+  resultsInnerLabel: DataResponse;
+};
 
-const ReadyMadeDonutChart = (props: ReadyMadeDonutChartProps) => {
+const DonutChartPro = (props: DonutLabelChartProProps) => {
   const theme = useTheme() as Theme;
+  const themeFormatter = getThemeFormatter(theme);
   i18nSetup(theme);
 
   const {
@@ -24,10 +31,24 @@ const ReadyMadeDonutChart = (props: ReadyMadeDonutChartProps) => {
     showTooltips,
     showValueLabels,
     title,
+    innerLabelMeasure,
+    resultsInnerLabel,
+    innerLabelText,
     onSegmentClick,
   } = resolveI18nProps(props);
 
   const data = getPieChartData({ data: results.data, dimension, measure, maxLegendItems }, theme);
+
+  const handleSegmentClick = (index: number | undefined) => {
+    onSegmentClick({
+      dimensionValue: index === undefined ? undefined : results.data?.[index]?.[dimension.name],
+    });
+  };
+
+  const label = themeFormatter.data(
+    innerLabelMeasure,
+    resultsInnerLabel?.data?.[0]?.[innerLabelMeasure.name],
+  );
 
   const options = mergician(
     getDefaultPieChartOptions(
@@ -39,14 +60,8 @@ const ReadyMadeDonutChart = (props: ReadyMadeDonutChartProps) => {
       } as DefaultPieChartOptions,
       theme,
     ),
-    theme.charts.donutChartOverrides ?? {},
+    theme.charts.donutLabelChartOverrides ?? {},
   );
-
-  const handleSegmentClick = (index: number | undefined) => {
-    onSegmentClick({
-      dimensionValue: index === undefined ? undefined : results.data?.[index]?.[dimension.name],
-    });
-  };
 
   return (
     <ChartCard
@@ -56,9 +71,15 @@ const ReadyMadeDonutChart = (props: ReadyMadeDonutChartProps) => {
       subtitle={description}
       title={title}
     >
-      <DonutChart data={data} options={options} onSegmentClick={handleSegmentClick} />
+      <DonutChart
+        label={label}
+        subLabel={innerLabelText}
+        data={data}
+        options={options}
+        onSegmentClick={handleSegmentClick}
+      />
     </ChartCard>
   );
 };
 
-export default ReadyMadeDonutChart;
+export default DonutChartPro;
