+++ src/remarkable-pro/components/component.constants.ts
