+++ src/remarkable-pro/theme/i18n/i18n.ts
