+++ src/remarkable-pro/theme/i18n/translations/de.ts
